Move ui folder contents to components
