+++ src/app/index.tsx
@@ -1,6 +1,6 @@
-import { UIProvider } from 'ui/UIProvider';
-import { Box } from 'ui/Box';
-import { TimerProvider } from 'ui/Timer';
+import { UIProvider } from 'components/UIProvider';
+import { Box } from 'components/Box';
+import { TimerProvider } from 'components/Timer';
 import { useThemedFavicon, useIntervalSequence } from 'hooks';
 import { ControlPanel } from 'components/ControlPanel';
 import { IntervalChip } from 'components/IntervalChip';
