+++ src/components/Box/types.ts
