+++ src/components/ControlPanel/index.tsx
@@ -1,8 +1,13 @@
 import { useEffect, useMemo, useRef, useState } from 'react';
 import { shallowEqual } from 'react-redux';
 import { Settings } from 'components/Settings';
-import { FinishTimer, TimeLeft, ToggleTimer, useTimerContext } from 'ui/Timer';
-import { Box } from 'ui/Box';
+import {
+  FinishTimer,
+  TimeLeft,
+  ToggleTimer,
+  useTimerContext,
+} from 'components/Timer';
+import { Box } from 'components/Box';
 import { useAppSelector } from 'store';
 import { selectCurrentInterval, selectIntervals } from 'store/intervals';
 import { selectIsSoundEnabled } from 'store/settings';
