+++ src/components/Icon/static/Break.tsx
