+++ src/components/Icon/static/Close.tsx
