+++ src/components/Icon/static/Dots.tsx
