+++ src/components/Icon/static/Focus.tsx
