+++ src/components/Icon/static/Forward.tsx
