+++ src/components/Icon/static/Pause.tsx
