+++ src/components/Icon/static/Play.tsx
