+++ src/components/Icon/types.ts
