+++ src/components/Settings/SettingsModal/SettingsForm.tsx
@@ -1,9 +1,9 @@
 import { ChangeEvent, useMemo } from 'react';
-import { TextField } from 'ui/TextField';
+import { TextField } from 'components/TextField';
 import { PomodoroIntervals } from 'types';
-import { Box } from 'ui/Box';
-import { Typography } from 'ui/Typography';
-import { Switch } from 'ui/Switch';
+import { Box } from 'components/Box';
+import { Typography } from 'components/Typography';
+import { Switch } from 'components/Switch';
 import { useAppSelector } from 'store';
 import { selectTheme } from 'store/theme';
 import { selectIsSoundEnabled } from 'store/settings';
