+++ src/components/Settings/SettingsModal/index.tsx
@@ -4,7 +4,7 @@ import { useAppSelector } from 'store';
 import { selectIntervals, selectPomodorosInSession } from 'store/intervals';
 import { useSettings } from 'hooks';
 import { PomodoroIntervals } from 'types';
-import { Modal } from 'ui/Modal';
+import { Modal } from 'components/Modal';
 import { SettingsForm } from './SettingsForm';
 
 type Props = {
