+++ src/components/Switch/Switch.styles.ts
