+++ src/components/Timer/FinishTimer.tsx
@@ -1,5 +1,5 @@
 import { useCallback } from 'react';
-import { Button } from 'ui/Button';
+import { Button } from 'components/Button';
 import { useTimerContext } from './TimerContext';
 
 export const FinishTimer = () => {
