+++ src/components/Timer/index.tsx
